requirements.txt
Flask==2.0.1
python-barcode==0.14.0
Pillow==9.5.0
numpy==1.23.5
pandas==1.5.3
Werkzeug==2.0.3
flask-cors==4.0.0
